backend/pyproject.toml
[project]
name = "auth-test-backend"
version = "0.1.0"
description = "FastAPI backend for authentication system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.6.0",
    "alembic>=1.13.0",
    "email-validator>=2.0.0",
    "passlib>=1.7.0",
    "bcrypt>=4.0.0,<5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9,<0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
